#4 - 엔티티 클래스 1차 생성
BaseEntity / 카테고리 / 장소 엔티티만 1차 생성
