--- src/main/java/com/dateplanner/constrant/BaseEntity.java
@@ -1,4 +1,0 @@
-package com.dateplanner.constrant;
-
-public class BaseEntity {
-}
