+++ src/main/java/com/dateplanner/place/entity/Category.java
@@ -1,20 +1,19 @@
-package com.dateplanner.place.domain;
+package com.dateplanner.place.entity;
 
+import com.dateplanner.constrant.entity.BaseEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Category {
+public class Category extends BaseEntity {
 
     @Id
     private String id;
